Play midi file

Play midi from MidiFramer object. Transform the midi file to fileext format and provide a play button for html documents.

WARNING: Setting overwrite = TRUE or marie_kondo = TRUE (the default!!) will DELETE the specified audio files!!! (see more details below)

player(
  midifile,
  soundfont = fluidsynth::soundfont_path(),
  output = gsub(".mid$", ".mp3", midifile),
  live = interactive(),
  verbose = interactive(),
  overwrite = TRUE,
  marie_kondo = TRUE,
  ...
)

Arguments

midifile

Path to the midi file to synthesize on; character string.

soundfont

path to sf2 sound font (character string); if not specified, the default soundfont of the fluidsynth package (fluidsynth::soundfont_path()) will be (downloaded if not present and) used.

output

Path to the audiofile to be synthesized. If audiofile of type mp3, it will first be synthesized to wav, and then converted to mp3 the av package; (character string).

live

logical; if TRUE the synthesized midi is directly played in the console. If FALSE an audio html tag is written. This will generate a small audio player when knitting an Rmd document (and probably also Quarto qmd files; I didn't check).

verbose

logical whether to print command line output; defaults to FALSE

overwrite

logical; defaults to TRUE; if file exists and overwrite = FALSE, the existing files will not be overwritten and the function errors out.

marie_kondo

logical, whether to remove intermediate files (the midi file, and if live = FALSE and audiofile is an mp3 also it also deletes the intermediate wav file); defaults to TRUE

...

Arguments passed to the fluidsynth functions (fluidsynth::midi_play or fluidsynth::midi_convert depending on the value of live).

Value

If live = TRUE, nothing is returned. If live = FALSE, a html audio tag is returned that will render as a small audio player when knitting an Rmd document. The audio player can then play the generated output audio file.

Examples

midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
midi_file_string |> player()
#> <audio controls="">
#>   <source src="/home/runner/work/_temp/Library/pyramidi/extdata/test_midi_file.mp3" type="audio/mp3"/>
#> </audio>

# The player is a small helper function to do basically this:
if (FALSE) {
midifile <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
audiofile <- "test.wav"
mp3file <- "test.mp3"
fluidsynth::midi_convert(midifile, output = audiofile)
av::av_audio_convert(audiofile, mp3file)
# `marie_kondo` = TRUE, the wav file would be deleted.
# `overwrite` = TRUE would delete the wav file
}