Tabulate measure related data in the midi event data

Tabulate measure related data in the midi event data

tab_measures(df, ticks_per_beat)

Arguments

df
Note event data.frame resulting of miditapyr$tidy_df()

ticks_per_beat
integer resulting of miditapyr$mido_midi_df()

Value

A data.frame with the following columns added: c("m", "b", "t", "ticks", "time").
m: measure
b: beat
t: time in seconds

Examples

## test
mt <- reticulate::import("miditapyr")

#> Error: Error 1 occurred creating conda environment r-reticulate

mido <- reticulate::import("mido")

#> Error: Error 1 occurred creating conda environment r-reticulate

mid_file_str <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido_mid_file <- mido$MidiFile(mid_file_str)

#> Error in eval(expr, envir, enclos): object 'mido' not found

dfc <- mt$midi_to_df(mido_mid_file)

#> Error in eval(expr, envir, enclos): object 'mt' not found

ticks_per_beat = mido_mid_file$ticks_per_beat

#> Error in eval(expr, envir, enclos): object 'mido_mid_file' not found

df <- dfc %>%
  mt$tidy_df()

#> Error in eval(lhs, parent, parent): object 'dfc' not found