Construct a midi frame R6 object

Construct a midi frame R6 object

Construct a midi frame R6 object

Value

MidiFramer R6 object

Details

Constructor of an object of class "MidiFramer".

Public fields

midi_file_string

Path to the midi file

mf

miditapyr$MidiFrames object,

dfm

result of tab_measures()

df_notes_long

result of pivot_long_notes()

df_meta, df_not_notes, df_notes_wide

results of triage_measured_unnested()

df_long_mod

result of merge_long_events()

Active bindings

df_meta, df_not_notes, df_notes_wide

results of triage_measured_unnested()

Methods

Public methods

MidiFramer$new()

MidiFramer$update_notes_wide()

MidiFramer$play()

MidiFramer$clone()

Method new()

Initialize a MidiFramer object

Usage

MidiFramer$new(midi_file_string)

Arguments

midi_file_string

Path to the midi file

Method update_notes_wide()

Update a MidiFramer object with modified notes

Usage

MidiFramer$update_notes_wide(mod)

Arguments

mod

Dataframe or function returning a dataframe of the format of df_notes_wide.

Examples

\dontrun{
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mfr <- MidiFramer$new(midi_file_string)
# Function to replace every note with a random midi note between 60 & 71:
mod <- function(dfn) {
  n_notes <- sum(!is.na(dfn$note))
  dfn %>% dplyr::mutate(note = ifelse(
    !is.na(note),
    sample(60:71, n_notes, TRUE),
    note
  ))
}
set.seed(123)
mfr$update_notes_wide(mod)
mfr$play()
# You can pass functions to the $update_notes_wide() method (as above), but
# you can also modify the dataframe directly and pass it. Therefore, the
# following results in the same:
set.seed(123)
df_mod <- mod(mfr$df_notes_wide)
mfr$update_notes_wide(df_mod)
mfr$play()
}

Method play()

Play midi from MidiFramer object Transform the midi file to fileext format and provide a play button for html documents.

Usage

MidiFramer$play(
  audiofile = tempfile("mf_out_", fileext = ".wav"),
  overwrite = FALSE
)

Arguments

audiofile

Path to the audiofile to be synthesized

overwrite

logical; defaults to FALSE; if file exists and overwrite = FALSE, the existing files will be used and nothing is synthesized/converted to audio files.

Examples

\dontrun{
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mfr <- MidiFramer$new(midi_file_string)
mfr$play()
}

Method clone()

The objects of this class are cloneable with this method.

Usage

MidiFramer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
MidiFramer$new(midi_file_string)
}

## ------------------------------------------------
## Method `MidiFramer$update_notes_wide`
## ------------------------------------------------

if (FALSE) {
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mfr <- MidiFramer$new(midi_file_string)
# Function to replace every note with a random midi note between 60 & 71:
mod <- function(dfn) {
  n_notes <- sum(!is.na(dfn$note))
  dfn %>% dplyr::mutate(note = ifelse(
    !is.na(note),
    sample(60:71, n_notes, TRUE),
    note
  ))
}
set.seed(123)
mfr$update_notes_wide(mod)
mfr$play()
# You can pass functions to the $update_notes_wide() method (as above), but
# you can also modify the dataframe directly and pass it. Therefore, the
# following results in the same:
set.seed(123)
df_mod <- mod(mfr$df_notes_wide)
mfr$update_notes_wide(df_mod)
mfr$play()
}

## ------------------------------------------------
## Method `MidiFramer$play`
## ------------------------------------------------

if (FALSE) {
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mfr <- MidiFramer$new(midi_file_string)
mfr$play()
}