MidiFramer class

The class MidiFramer can be used to read midi files to dataframes in order to facilitate to manipulate the data from R. You can also create midi data from R without reading it from a file. The data is transformed to various formats. One of the MidiFramer fields is a MidiFrames object of the python miditapyr package. Its method write_file() can be used to write the data back to a midi file.

Details

See the vignette("midi_framer") for a brief usage introduction how to manipulate midi data.

The vignette("compose") shows a more extended example how to generate midi files from scratch.

vignette("package_workflow") shows in detail the structure of the MidiFramer class.

vignette("functions_usage") illustrates the low-level functions of the pyramidi package. that MidiFramer objects use under the hood.

Public fields

midi_file_string

Path to the midi file.

mf

miditapyr$MidiFrames object.

dfm

result of tab_measures().

df_notes_long

Result of pivot_long_notes().

df_meta, df_not_notes, df_notes_wide

Results of split_midi_frame().

midi_frame_mod

Result of merge_midi_frames().

params

Parameters used in internal functions; Named list; params$columns_to_add is passed to tab_measures(columns_to_add).

Active bindings

df_meta, df_not_notes, df_notes_wide

Results of split_midi_frame().

ticks_per_beat

Set ticks per beat of MidiFrames()$mf$midi_file. The value of ticks_per_beat passed should be integer. When a value is passed, the field mf$midi_file field$ticks_per_beat is modified.

Methods

Public methods

MidiFramer$new()

MidiFramer$update_notes_wide()

MidiFramer$populate_r_fields()

MidiFramer$play()

MidiFramer$clone()

Method new()

Initialize a MidiFramer object

Usage

MidiFramer$new(midi_file_string = NULL)

Arguments

midi_file_string

Path to the midi file; if NULL (the default), an empty MidiFramer object is created.

Method update_notes_wide()

Update a MidiFramer object with modified notes

Usage

MidiFramer$update_notes_wide(mod)

Arguments

mod

Dataframe or function returning a dataframe of the format of df_notes_wide.

Examples

\dontrun{
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mfr <- MidiFramer$new(midi_file_string)
# Function to replace every note with a random midi note between 60 & 71:
mod <- function(dfn) {
  n_notes <- sum(!is.na(dfn$note))
  dfn %>% dplyr::mutate(note = ifelse(
    !is.na(note),
    sample(60:71, n_notes, TRUE),
    note
  ))
}
set.seed(123)
mfr$update_notes_wide(mod)
mfr$play()
# You can pass functions to the $update_notes_wide() method (as above), but
# you can also modify the dataframe directly and pass it. Therefore, the
# following results in the same:
set.seed(123)
df_mod <- mod(mfr$df_notes_wide)
mfr$update_notes_wide(df_mod)
mfr$play()
}

Method populate_r_fields()

Populate the fields of a MidiFramer object

This can also be used to recalculate all the object's attributes, when a value in params is changed (see examples).

Usage

MidiFramer$populate_r_fields()

Examples

\dontrun{
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mfr <- MidiFramer$new(midi_file_string)
mfr$params$columns_to_add <- c("m", "b", "t", "time")
mfr$populate_r_fields()
}

Method play()

Play midi from MidiFramer object. Transform the midi file to fileext format and provide a play button for html documents.

Usage

MidiFramer$play(
  audiofile = tempfile("mf_out_", fileext = ".wav"),
  overwrite = FALSE,
  soundfont = NULL,
  verbose = FALSE
)

Arguments

audiofile

Path to the audiofile to be synthesized. If audiofile of type mp3, it will first be synthesized to wav, and then converted to mp3 with ffmpeg.

overwrite

logical; defaults to FALSE; if file exists and overwrite = FALSE, the existing files will be used and nothing is synthesized/converted to audio files.

soundfont

path to sf2 sound font (character string); if NULL (the default) the default soundfont is used.

verbose

logical whether to print fluidsynth (and ffmpeg in case of mp3 audiofile type) command line output; defaults to FALSE

Examples

\dontrun{
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mfr <- MidiFramer$new(midi_file_string)
mfr$play()
}

Method clone()

The objects of this class are cloneable with this method.

Usage

MidiFramer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
## Create a MidiFramer object from a midi file:
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
MidiFramer$new(midi_file_string)

## ------------------------------------------------
## Create empty MidiFramer object to illustrate
## the use of the `ticks_per_beat` active binding:
## ------------------------------------------------

mfr <- MidiFramer$new()
# Print default value of empty MidiFile object:
mfr$mf$midi_file$ticks_per_beat
# Modify it with the active binding ticks_per_beat:
mfr$ticks_per_beat <- 960L
# Print it again:
mfr$mf$midi_file$ticks_per_beat
}

## ------------------------------------------------
## Method `MidiFramer$update_notes_wide`
## ------------------------------------------------

if (FALSE) {
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mfr <- MidiFramer$new(midi_file_string)
# Function to replace every note with a random midi note between 60 & 71:
mod <- function(dfn) {
  n_notes <- sum(!is.na(dfn$note))
  dfn %>% dplyr::mutate(note = ifelse(
    !is.na(note),
    sample(60:71, n_notes, TRUE),
    note
  ))
}
set.seed(123)
mfr$update_notes_wide(mod)
mfr$play()
# You can pass functions to the $update_notes_wide() method (as above), but
# you can also modify the dataframe directly and pass it. Therefore, the
# following results in the same:
set.seed(123)
df_mod <- mod(mfr$df_notes_wide)
mfr$update_notes_wide(df_mod)
mfr$play()
}

## ------------------------------------------------
## Method `MidiFramer$populate_r_fields`
## ------------------------------------------------

if (FALSE) {
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mfr <- MidiFramer$new(midi_file_string)
mfr$params$columns_to_add <- c("m", "b", "t", "time")
mfr$populate_r_fields()
}

## ------------------------------------------------
## Method `MidiFramer$play`
## ------------------------------------------------

if (FALSE) {
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mfr <- MidiFramer$new(midi_file_string)
mfr$play()
}